@@ -1,8 +1,8 @@
 package dev.guipalazzo.spring.api.domain;
 
+import jakarta.persistence.*;
 import lombok.*;
 
-import javax.persistence.*;
 import java.math.BigDecimal;
 import java.util.List;
 
