@@ -8,6 +8,6 @@ public enum FormaPagamento {
     CARTAO_CREDITO,
     CARTAO_DEBITO,
     PIX,
-    DINHEIRO;
+    DINHEIRO
 
 }
