@@ -8,5 +8,5 @@ public enum StatusPagamento {
     PENDENTE,
     PAGO,
     ESTORNADO,
-    CANCELADO;
+    CANCELADO
 }
