@@ -3,6 +3,6 @@ package dev.guipalazzo.spring.api.domain;
 public enum TipoAnuncio {
 
     COMPLETO,
-    QUARTO;
+    QUARTO
 
 }
